+++ KSP/src/ksp/sandeliavimas/PaskirstymasController.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package ksp;
+package ksp.sandeliavimas;
 
 import java.net.URL;
 import java.util.ResourceBundle;
@@ -12,14 +12,15 @@ import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
-import javafx.scene.control.TextField;
+import ksp.LoginManager;
+import ksp.MainViewManager;
 
 /**
  * FXML Controller class
  *
  * @author Martynas
  */
-public class PridetiPrekeController implements Initializable {
+public class PaskirstymasController implements Initializable {
 
     @FXML
     private Button logoutButton;
@@ -30,19 +31,7 @@ public class PridetiPrekeController implements Initializable {
     @FXML
     private Button backButton;
     @FXML
-    private Button addToWarehouse;
-    @FXML
-    private TextField name;
-    @FXML
-    private TextField code;
-    @FXML
-    private TextField brand;
-    @FXML
-    private TextField country;
-    @FXML
-    private TextField wight;
-    @FXML
-    private Button add;
+    private Button distribute;
 
     /**
      * Initializes the controller class.
@@ -62,9 +51,6 @@ public class PridetiPrekeController implements Initializable {
         });
         backButton.setOnAction((ActionEvent event) -> {
            mainManager.navigateStorage(loginManager, sessionID);
-        });
-        addToWarehouse.setOnAction((ActionEvent event) -> {
-            sandeliavimas.navigateAddToWareHouse(mainManager, loginManager, sessionID);
         });
     }
