+++ KSP/src/ksp/sandeliavimas/PrekiuSalinimasController.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package ksp;
+package ksp.sandeliavimas;
 
 import java.net.URL;
 import java.util.ResourceBundle;
@@ -13,13 +13,15 @@ import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.control.TextField;
+import ksp.LoginManager;
+import ksp.MainViewManager;
 
 /**
  * FXML Controller class
  *
  * @author Martynas
  */
-public class PridetiISandeliController implements Initializable {
+public class PrekiuSalinimasController implements Initializable {
 
     @FXML
     private Button logoutButton;
@@ -30,13 +32,9 @@ public class PridetiISandeliController implements Initializable {
     @FXML
     private Button backButton;
     @FXML
-    private Button addNew;
-    @FXML
-    private TextField price;
-    @FXML
     private TextField count;
     @FXML
-    private Button add;
+    private Button delete;
 
     /**
      * Initializes the controller class.
@@ -56,9 +54,6 @@ public class PridetiISandeliController implements Initializable {
         });
         backButton.setOnAction((ActionEvent event) -> {
            mainManager.navigateStorage(loginManager, sessionID);
-        });
-        addNew.setOnAction((ActionEvent event) -> {
-            sandeliavimas.navigateAdd(mainManager, loginManager, sessionID);
         });
     }
