+++ KSP/src/ksp/sandeliavimas/SandeliavimasController.java
@@ -1,8 +1,10 @@
-package ksp;
+package ksp.sandeliavimas;
 
 import javafx.event.*;
 import javafx.fxml.FXML;
 import javafx.scene.control.*;
+import ksp.LoginManager;
+import ksp.MainViewManager;
 
 /** Controls the main application screen */
 public class SandeliavimasController {
