+++ KSP/src/ksp/sandeliavimas/pridetiPreke.fxml
